+++ README.md
@@ -1,4 +1,13 @@
 # VirtualMachine-task-FAT-4357-
+
 internship tasks
 
-fix patterns and dependencies
+---
+
+## fix patterns and dependencies
+
+division into modules
+working with webpack
+install pretier
+
+---
