+++ prettier.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  printWidth: 80,
+  singleQuote: true,
+  trailingComma: 'all',
+  bracketSpacing: true,
+  jsxBracketSameLine: false,
+  tabWidth: 5,
+  semi: true,
+};
