+++ src/index.html
@@ -25,7 +25,7 @@
         
     </div>
 
-    <script src="index.js"></script>
+    <script src="../dist/bundle.js"></script>
 
 </body>
