+++ webpack.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  entry: './src/app.js',
+  output: {
+    filename: 'bundle.js',
+    path: __dirname + '/dist',
+  },
+};
